Refactor implementation of fragments.
Fragments deal with how content is shown piece by piece.  For example, if we
have the slide:

    Foo

    . . .

    Bar

There are two fragments.  First, only `Foo` is visible.  When the user goes
to the next fragment, `Foo` and `Bar` are both visible.

The first implementation of fragments in patat would compute the state of
all the visible items.  In this case, that would be:

1.  `[Foo]`
2.  `[Foo, Bar]`

This does not work elegantly if you want another pass to add further fragments:
you now need to split `Foo` in multiple places (and the thunk is no longer
shared).  This PR refactors this to use "instructions" over the slide content.
For the example, these instructions would be:

1.  `Append [Foo]`
2.  `Pause`
3.  `Append [Bar]`

The `Pause`s are explicit, and indicate how many fragments are present.

There are more constructors, for manipulation of lists which is necessary
if nested lists are shown incrementally.  This all for allows much nicer
manipulation of the fragments, which in turn is useful for #52.