Allow overriding certain settings in slides (#148)
Configuration was typically done in the metadata block of the input file,
or in a per-user configuration.  These settings are applied to the entire
presentation.

We now allow selectively overriding these settings on a per-slide basis, by
adding one or more config blocks to those slides.  Config blocks are comments
that start with `config:`.  They can be placed anywhere in the slide.

    # This is a normal slide

    Normal slide content

    # This slide has a different colour header

    <!--config:
    theme:
      header: [vividYellow]
    -->

    Wow, how did that happen?