rename eval.wrap to eval.container (#167)
`wrap` is used at the top-level of settings for wrapping at a certain column, and inside `eval` to determine the type in which the result is "wrapped". Using the same name for both is confusing, so this adds `eval.container` as the new name for `eval.wrap`.  `eval.wrap` will continue to be supported for the forseeable future, but its use will be discouraged.

This also changes the values (again keeping the original ones for backwards-compat), so the complete changes to a configuration would be:

- `wrap: code` -> `container: code`
- `wrap: raw` -> `container: none`
- `wrap: rawInline` -> `container: inline`